+++ src/app/cli.zig
@@ -15,7 +15,7 @@ pub const Run = struct {
     cfg: config.Config,
     session: Args.SessionSel = .auto,
     no_session: bool = false,
-    tool_mask: u8 = @import("../core/mod.zig").tools.builtin.mask_all,
+    tool_mask: u16 = @import("../core/mod.zig").tools.builtin.mask_all,
     thinking: Args.ThinkingLevel = .adaptive,
     verbose: bool = false,
     system_prompt: ?[]const u8 = null,
